# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zenith Staff Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/discord/callback:
    get:
      operationId: discordAuthCallback
      tags: [auth]
      summary: Discord OAuth callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to dashboard
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/guilds:
    get:
      operationId: getUserGuilds
      tags: [auth]
      summary: Get guilds the user manages
      responses:
        "200":
          description: List of manageable guilds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscordGuild"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /guilds/{guildId}:
    get:
      operationId: getGuild
      tags: [guilds]
      summary: Get guild configuration
      parameters:
        - $ref: "#/components/parameters/GuildId"
      responses:
        "200":
          description: Guild config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guild"
        "404":
          description: Guild not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateGuild
      tags: [guilds]
      summary: Update guild configuration
      parameters:
        - $ref: "#/components/parameters/GuildId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuildBody"
      responses:
        "200":
          description: Updated guild
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guild"
  /guilds/{guildId}/stats:
    get:
      operationId: getGuildStats
      tags: [stats]
      summary: Get dashboard statistics for a guild
      parameters:
        - $ref: "#/components/parameters/GuildId"
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildStats"
...
  /guilds/{guildId}/divisions:
    get:
      operationId: listDivisions
      tags: [divisions]
      summary: List divisions
      parameters:
        - $ref: "#/components/parameters/GuildId"
      responses:
        "200":
          description: Divisions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Division"
    post:
      operationId: createDivision
      tags: [divisions]
      summary: Create a division
      parameters:
        - $ref: "#/components/parameters/GuildId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDivisionBody"
      responses:
        "201":
          description: Division created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Division"
  /guilds/{guildId}/divisions/{divisionId}:
    patch:
      operationId: updateDivision
      tags: [divisions]
      summary: Update a division
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: divisionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDivisionBody"
      responses:
        "200":
          description: Updated division
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Division"
    delete:
      operationId: deleteDivision
      tags: [divisions]
      summary: Delete a division
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: divisionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /guilds/{guildId}/applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResponse"
  /guilds/{guildId}/applications/{applicationId}/accept:
    post:
      operationId: acceptApplication
      tags: [applications]
      summary: Accept an application
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewApplicationBody"
      responses:
        "200":
          description: Application accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /guilds/{guildId}/applications/{applicationId}/deny:
    post:
      operationId: denyApplication
      tags: [applications]
      summary: Deny an application
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewApplicationBody"
      responses:
        "200":
          description: Application denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /guilds/{guildId}/questions:
    get:
      operationId: listApplicationQuestions
      tags: [applications]
      summary: List application questions
      parameters:
        - $ref: "#/components/parameters/GuildId"
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationQuestion"
    post:
      operationId: createApplicationQuestion
      tags: [applications]
      summary: Create application question
      parameters:
        - $ref: "#/components/parameters/GuildId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationQuestionBody"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationQuestion"
  /guilds/{guildId}/questions/{questionId}:
    patch:
      operationId: updateApplicationQuestion
      tags: [applications]
      summary: Update application question
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationQuestion"
    delete:
      operationId: deleteApplicationQuestion
      tags: [applications]
      summary: Delete application question
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    GuildId:
      name: guildId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    DiscordUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        discriminator:
          type: string
        avatar:
          type: string
          nullable: true
        avatarUrl:
          type: string
      required: [id, username, discriminator, avatarUrl]
    DiscordGuild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
          nullable: true
        iconUrl:
          type: string
          nullable: true
        isSetup:
          type: boolean
      required: [id, name, isSetup]
    Guild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        iconUrl:
          type: string
          nullable: true
        isPremium:
          type: boolean
        staffRoleId:
          type: string
          nullable: true
        managementRoleId:
          type: string
          nullable: true
        logChannelId:
          type: string
          nullable: true
        applicationChannelId:
          type: string
          nullable: true
        applicationReviewChannelId:
          type: string
          nullable: true
        welcomeChannelId:
          type: string
          nullable: true
        embedColor:
          type: string
        embedFooter:
          type: string
          nullable: true
        customBotName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, isPremium, embedColor, createdAt]
    UpdateGuildBody:
      type: object
      properties:
        staffRoleId:
          type: string
          nullable: true
        managementRoleId:
          type: string
          nullable: true
        logChannelId:
          type: string
          nullable: true
        applicationChannelId:
          type: string
          nullable: true
        applicationReviewChannelId:
          type: string
          nullable: true
        welcomeChannelId:
          type: string
          nullable: true
        embedColor:
          type: string
        embedFooter:
          type: string
          nullable: true
        customBotName:
          type: string
          nullable: true
        customBotAvatar:
          type: string
          nullable: true
    GuildStats:
      type: object
      properties:
        totalStaff:
          type: integer
        activeStaff:
          type: integer
        pendingApplications:
          type: integer
        activeStrikes:
          type: integer
        activeLoas:
          type: integer
        totalPromotions:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [totalStaff, activeStaff, pendingApplications, activeStrikes, activeLoas, totalPromotions, recentActivity]
    ActivityItem:
      type: object
      properties:
        type:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
        actorUsername:
          type: string
      required: [type, description, timestamp, actorUsername]
...
    Division:
      type: object
      properties:
        id:
          type: string
        guildId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        discordRoleId:
          type: string
          nullable: true
        channelId:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, guildId, name, isActive, createdAt]
    CreateDivisionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        discordRoleId:
          type: string
          nullable: true
        channelId:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
      required: [name]
    UpdateDivisionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        discordRoleId:
          type: string
          nullable: true
        channelId:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        isActive:
          type: boolean
    Application:
      type: object
      properties:
        id:
          type: string
        guildId:
          type: string
        applicantDiscordId:
          type: string
        applicantUsername:
          type: string
        robloxUsername:
          type: string
          nullable: true
        status:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationAnswer"
        reviewerId:
          type: string
          nullable: true
        reviewerUsername:
          type: string
          nullable: true
        reviewNotes:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, guildId, applicantDiscordId, applicantUsername, status, answers, submittedAt]
    ApplicationAnswer:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required: [question, answer]
    ApplicationListResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [applications, total, page, totalPages]
    ReviewApplicationBody:
      type: object
      properties:
        notes:
          type: string
          nullable: true
    ApplicationQuestion:
      type: object
      properties:
        id:
          type: string
        guildId:
          type: string
        question:
          type: string
        placeholder:
          type: string
          nullable: true
        isRequired:
          type: boolean
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, guildId, question, isRequired, position, createdAt]
    CreateApplicationQuestionBody:
      type: object
      properties:
        question:
          type: string
        placeholder:
          type: string
          nullable: true
        isRequired:
          type: boolean
        position:
          type: integer
      required: [question]
    UpdateApplicationQuestionBody:
      type: object
      properties:
        question:
          type: string
        placeholder:
          type: string
          nullable: true
        isRequired:
          type: boolean
        position:
          type: integer
...